Add item info

+++ accounting_bot/ext/echoes_data.py
@@ -4,29 +4,25 @@
 # Load-After: [accounting_bot.ext.sheet.projects, accounting_bot.ext.sheet.sheet_main]
 # End
 import asyncio
-import csv
 import difflib
-import io
 import logging
 import os
-from asyncore import write
 from pathlib import Path
-from typing import Dict, Any, List, Optional, TYPE_CHECKING
+from typing import Dict, Any, List, Optional, TYPE_CHECKING, Union
 
 import discord
-from blue_echoes.data import Item
+from blue_echoes.data import Item, CompressionType, is_loaded
 from blue_echoes.data.models import EstimatedMarketData
 from blue_echoes.data_utils import Dialect
-from discord import SlashCommandGroup, option, ApplicationContext, Embed
+from discord import SlashCommandGroup, option, ApplicationContext, Embed, Message, Webhook
 from discord.ext import commands
 from gspread.utils import ValueInputOption
 from sqlalchemy.ext.asyncio import create_async_engine
 
-from accounting_bot import utils
 from accounting_bot.exceptions import PluginNotFoundException
 from accounting_bot.main_bot import BotPlugin, AccountingBot, PluginWrapper, PluginState
 from accounting_bot.utils import AutoDisableView
-from accounting_bot.utils.ui import NumPadView, ModalForm, ConfirmView, AwaitConfirmView
+from accounting_bot.utils.ui import ModalForm, AwaitConfirmView
 
 if TYPE_CHECKING:
     from accounting_bot.ext.sheet.sheet_main import SheetPlugin
@@ -116,6 +112,14 @@ def build_cost_formulars(item: Item, atomic_resources: List[str], automatic_reso
     return rows
 
 
+display_attrs = [
+    (1, "Tech Level", "T{val:0.0f}"),
+    (2, "Ship Size", "{val:0.0f}"),
+    (4, "Meta Level", "{val:0.0f}"),
+    (13, "Module Size", "{val:0.0f}"),
+]
+
+
 class MarketGroupSelect(discord.ui.Select):
     def __init__(
             self,
@@ -126,20 +130,47 @@ class MarketGroupSelect(discord.ui.Select):
             placeholder="Select a market group",
             **kwargs
     ):
-        self.select_options = {group["id"]: discord.SelectOption(label=group["name"], value=str(group["id"])) for group in
-                   select_options.values()}
+        self.select_options = {group["id"]: discord.SelectOption(label=group["name"], value=str(group["id"])) for group
+                               in
+                               select_options.values()}
+        self._prev_page_option = discord.SelectOption(label="Previous Page", value="prev", emoji="◀")
+        self._next_page_option = discord.SelectOption(label="Next Page", value="next", emoji="▶")
         self._callback = callback
         self.parent_id = parent_id
+        self.page = 0
+        self._is_paged = len(self.select_options) > 25
         super().__init__(
             placeholder=placeholder,
             min_values=1,
             max_values=1,
-            options=list(self.select_options.values()),
             *args, **kwargs
         )
+        self.reset_options()
+
+    def reset_options(self):
+        self._is_paged = len(self.select_options) > 25
+        if self._is_paged:
+            self.options = [*list(self.select_options.values())[:23], self._next_page_option]
+        else:
+            self.options = list(self.select_options.values())
 
     async def callback(self, interaction: discord.Interaction):
-        await self._callback(self, interaction)
+        if self._is_paged and self.values[0] in ("prev", "next"):
+            if self.values[0] == "prev":
+                self.page -= 1
+            elif self.values[0] == "next":
+                self.page += 1
+            start = self.page * 23
+            end = start + 23
+            options = list(self.select_options.values())[start:end]
+            if start > 0:
+                options.insert(0, self._prev_page_option)
+            if end < len(self.select_options):
+                options.append(self._next_page_option)
+            self.options = options
+            await self._callback(self, interaction, new_page=True)
+        else:
+            await self._callback(self, interaction)
 
 
 class ItemSelectView(AutoDisableView):
@@ -148,14 +179,12 @@ class ItemSelectView(AutoDisableView):
                  market_groups: Dict[int, Dict[str, Any]],
                  callback,
                  auto_delete=True,
-                 silent=True,
                  *args, **kwargs,
                  ):
         super().__init__(*args, **kwargs)
         self.plugin = plugin
         self.market_groups = market_groups
         self.auto_delete = auto_delete
-        self.silent = silent
         primary_group = dict(filter(lambda x: x[1]["parent_id"] is None, market_groups.items()))
         # self.secondary_group = dict(filter(lambda x: x[1]["parent_id"] in self.primary_group, market_groups.items()))
         # self.tertiary_group = dict(filter(lambda x: x[1]["parent_id"] in self.secondary_group, market_groups.items()))
@@ -168,12 +197,16 @@ class ItemSelectView(AutoDisableView):
         self.market_selection: List[Optional[int]] = [None, None, None, None]
         self._callback = callback
 
-    async def select_callback(self, select, interaction: discord.Interaction):
+    async def select_callback(self, select, interaction: discord.Interaction, new_page=False):
+        if new_page:
+            await interaction.response.defer(invisible=True)
+            await self.message.edit(view=self)
+            return
         if select == self.market_drawers[3]:
             task = None
             if self.auto_delete:
                 task = asyncio.create_task(self.message.delete())
-            await self._callback(interaction, int(select.values[0]), silent=self.silent)
+            await self._callback(interaction, self.message, int(select.values[0]))
             if task is not None:
                 await task
             if not self.auto_delete:
@@ -241,24 +274,12 @@ class ItemSelectView(AutoDisableView):
             if self.market_selection[i] is not None:
                 opt = self.market_drawers[i].select_options[self.market_selection[i]]
                 opt.default = True
-            self.market_drawers[i].options = list(self.market_drawers[i].select_options.values())
-        # debug_msg = "Selection: " + ",".join(str(x) for x in self.market_selection) + "\n"
-        # for i in range(4):
-        #     if self.market_drawers[i] is None:
-        #         debug_msg += f"Drawer {i}: None\n"
-        #     else:
-        #         item = self.market_drawers[i]
-        #         debug_msg += f"Drawer {i}: Parent: {item.parent_id}, options: {len(item.select_options)}\n"
-        #         for opt in item.options:
-        #             if not opt.default:
-        #                 continue
-        #             debug_msg += f"  Selected: {opt.label} ({opt.value})\n"
+            self.market_drawers[i].reset_options()
         await self.message.edit(
             view=self)
 
 
 class BlueprintInfoView(AutoDisableView):
-
     def __init__(self, item: Item, plugin: EchoesDataPlugin, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.item = item
@@ -271,7 +292,8 @@ class BlueprintInfoView(AutoDisableView):
         prices = await self.plugin.db.market_repo.fetch_last_estimated_prices(item_ids)
         price_item: EstimatedMarketData = next(filter(lambda p: p.type_id == self.item.id, prices), None)
         if self.item.blueprint is not None:
-            price_bp: EstimatedMarketData = next(filter(lambda p: p.type_id == self.item.blueprint.blueprint_item.id, prices), None)
+            price_bp: EstimatedMarketData = next(
+                filter(lambda p: p.type_id == self.item.blueprint.blueprint_item.id, prices), None)
         else:
             price_bp = None
         embed_desc = f"Item ID: `{self.item.id}`\n"
@@ -281,7 +303,10 @@ class BlueprintInfoView(AutoDisableView):
             latest_time = price_item.week_of_patch
         if price_bp is not None:
             embed_desc += f"Blueprint Price: `{price_bp.average_price_no_outliers:,} ISK`\n"
-            latest_time = max(latest_time, price_bp.week_of_patch)
+            if latest_time is None:
+                latest_time = price_bp.week_of_patch
+            elif price_bp.week_of_patch is not None:
+                latest_time = max(latest_time, price_bp.week_of_patch)
         if latest_time is not None:
             embed_desc += f"-# Time of price data: <t:{int(latest_time.timestamp())}:f>\n"
         embed = Embed(
@@ -321,7 +346,11 @@ class BlueprintInfoView(AutoDisableView):
         interaction = model.get_interaction()
         await interaction.response.defer(ephemeral=True)
         sheet_name = result["Sheet Name"]
-        start_row = int(result["Row"].strip())
+        try:
+            start_row = int(result["Row"].strip())
+        except ValueError:
+            await interaction.followup.send(f"Invalid row number `{result['Row']}`", ephemeral=True)
+            return
         sheet_plugin: SheetPlugin = self.plugin.bot.get_plugin("SheetMain")
         sheet = await sheet_plugin.get_sheet()
         worksheets = await sheet.worksheets()
@@ -383,24 +412,185 @@ class EchoesDataCommands(commands.Cog):
 
     def __init__(self, plugin: EchoesDataPlugin):
         self.plugin = plugin
+        self._market_groups = None
+
+    @property
+    async def market_groups(self) -> Dict[int, Dict[str, Union[int, str, None]]]:
+        if self._market_groups is not None:
+            return self._market_groups
+        self._market_groups = await self.plugin.db.item_repo.fetch_market_groups()
+        return self._market_groups
 
     @group.command(name="blueprint", description="Tools for blueprint data")
     @option(name="item_name", description="The name of the item", type=str, required=False, default=None)
     @option(name="silent", description="Execute the command silently", type=bool, required=False, default=True)
     async def cmd_bp(self, ctx: ApplicationContext, item_name: str, silent: bool):
-        await ctx.defer(ephemeral=True)
-        market_groups = await self.plugin.db.item_repo.fetch_market_groups()
+        await ctx.defer(ephemeral=silent)
+        market_groups = await self.market_groups
+        if item_name is not None:
+            item = await self.plugin.db.item_repo.fetch_item(item_name=item_name)
+            await self.cmd_bp_callback(ctx, ctx.followup, item.id, item, defer_response=False, silent=silent)
+        else:
+            view = ItemSelectView(self.plugin, market_groups,
+                                  callback=self.cmd_bp_callback, auto_delete=False)
+            msg = await ctx.followup.send("Select an item", view=view)
+            view.message = msg
 
-        view = ItemSelectView(self.plugin, market_groups,
-                              callback=self.cmd_bp_callback, auto_delete=False, silent=silent)
-        msg = await ctx.followup.send("Select an item", view=view)
-        view.message = msg
-
-    async def cmd_bp_callback(self, interaction: discord.Interaction, item_id: int, silent=True):
-        await interaction.response.defer(ephemeral=silent)
-        item = await self.plugin.db.item_repo.fetch_item(item_id=item_id)
+    async def cmd_bp_callback(
+            self, interaction: discord.Interaction, message: Union[Message, Webhook],
+            item_id: int, item: Optional[Item] = None, defer_response=True, silent=True):
+        if defer_response:
+            await interaction.response.defer(ephemeral=True, invisible=True)
+        if item is None:
+            item = await self.plugin.db.item_repo.fetch_item(item_id=item_id)
         await self.plugin.db.item_repo.fetch_blueprint_data(item, recursive=True)
         view = BlueprintInfoView(item, self.plugin)
         embed = await view.build_embed()
-        msg = await interaction.followup.send(ephemeral=silent, view=view, embed=embed)
+        if isinstance(message, Message):
+            msg = await message.edit(content="", view=view, embed=embed)
+        elif isinstance(message, Webhook):
+            msg = await message.send(embed=embed, view=view, ephemeral=silent)
+        else:
+            raise TypeError(f"Invalid message type {type(message)}, expected Message or Webhook")
         view.message = msg
+
+    @group.command(name="item", description="Tools for blueprint data")
+    @option(name="item_name", description="The name of the item, use % as a wildcard", type=str, required=False,
+            default=None)
+    @option(name="silent", description="Execute the command silently", type=bool, required=False, default=True)
+    async def cmd_item(self, ctx: ApplicationContext, item_name: str, silent: bool):
+        await ctx.response.defer(ephemeral=silent)
+        market_groups = await self.market_groups
+        if item_name is None:
+            view = ItemSelectView(self.plugin, market_groups,
+                                  callback=self.cmd_item_callback, auto_delete=False)
+            msg = await ctx.followup.send("Select an item", view=view)
+            view.message = msg
+        else:
+            item = await self.plugin.db.item_repo.fetch_item(
+                item_name=item_name,
+                include_attributes=True,
+                include_compression=True,
+                include_reprocessing=True,
+                include_corp_task=True
+            )
+            await self.cmd_item_callback(None, ctx.followup, item.id, item, silent=silent)
+
+    async def cmd_item_callback(
+            self, interaction: discord.Interaction, message: Union[Message, Webhook],
+            item_id: int, item: Optional[Item] = None, silent=True
+    ):
+        if interaction is not None:
+            await interaction.response.defer(invisible=True)
+        if item is None:
+            item = await self.plugin.db.item_repo.fetch_item(
+                item_id=item_id,
+                include_attributes=True,
+                include_compression=True,
+                include_reprocessing=True,
+                include_corp_task=True
+            )
+        embed = await self.build_item_embed(item)
+        if isinstance(message, Message):
+            await message.edit(content="", embed=embed, view=None)
+        elif isinstance(message, Webhook):
+            await message.send(embed=embed, ephemeral=silent)
+        else:
+            raise TypeError(f"Invalid message type {type(message)}, expected Message or Webhook")
+
+    async def build_item_embed(self, item: Item):
+        market_groups = await self.market_groups
+        prices = await self.plugin.db.market_repo.fetch_last_estimated_prices([item.id])
+
+        content = ""
+        if item.market_group_id is not None:
+            tertiary_group = market_groups[item.market_group_id]
+            secondary_group = market_groups[tertiary_group["parent_id"]]
+            primary_group = market_groups[secondary_group["parent_id"]]
+            content += (f"-# {primary_group['name']} "
+                        f"> {secondary_group['name']} "
+                        f"> {tertiary_group['name']}\n")
+        content += f"Item ID: `{item.id}`\n"
+        if len(prices) == 1:
+            price_item = prices[0]
+            content += f"Price: `{price_item.average_price_no_outliers:,} ISK`\n"
+            content += f"-# Time of price data: <t:{int(price_item.week_of_patch.timestamp())}:f>\n"
+        embed = Embed(
+            title=f"Item Info for {item.name}",
+            description=content,
+            color=discord.Color.red()
+        )
+        content = ""
+        if is_loaded(item.blueprint) and item.blueprint is not None:
+            content += f"Blueprint: `{item.blueprint.blueprint_item.name}`\n"
+            content += f"Blueprint ID: `{item.blueprint.blueprint_item.id}`\n"
+            content += f"Station Price: `{item.blueprint.money:,} ISK`\n"
+            content += f"Time: `{item.blueprint.time}`\n"
+            content += f"Output num: `{item.blueprint.output_num}`\n"
+            embed.add_field(name="Blueprint", value=content, inline=False)
+
+        content = (f"Base Price: `{item.base_price:,} ISK`\n"
+                   f"Volume: `{item.volume:,} m³`\n"
+                   f"Can Be Jetted: `{item.can_be_jettisoned}`\n"
+                   f"Is Omega: `{item.is_omega}`\n"
+                   # f"Tradeable: `{item.contract_allowed}`\n"  # Doesn't work
+        )
+        if item.corp_task:
+            content += f"Corp Task: `{item.corp_task.fp_reward} FP`, Group: `{item.corp_task.random_group}`\n"
+        embed.add_field(
+            name="Properties",
+            value=content
+        )
+
+        if is_loaded(item.attributes) and len(item.attributes) > 0:
+            content = ""
+            for attr_id, attr_name, attr_unit in display_attrs:
+                attr = item.get_attribute_by_id(attr_id)
+                if attr is None or attr.value == 0.0:
+                    continue
+                content += f"{attr_name}: "
+                if attr_unit is None:
+                    content += f"`{attr.value}`"
+                else:
+                    content += f"`{attr_unit.format(val=attr.value)}`"
+                content += "\n"
+            embed.add_field(
+                name="Attributes",
+                value=content
+            )
+        if item.main_cal_code is not None or item.online_cal_code is not None or item.active_cal_code is not None:
+            embed.add_field(
+                name="Modifiers",
+                value=f"Main: `{item.main_cal_code}`\n"
+                      f"Online: `{item.online_cal_code}`\n"
+                      f"Active: `{item.active_cal_code}`",
+                inline=False
+            )
+        if is_loaded(item.reprocessing) and len(item.reprocessing) > 0:
+            content = "-# With Skill 000\n```\n"
+            max_len = max([len(rep.result.name) for rep in item.reprocessing])
+            if max_len > 25:
+                max_len = min(max_len, 35)
+            for rep in item.reprocessing:
+                content += f"{rep.result.name: <{max_len}}: {rep.quantity:,}\n"
+            content += "```"
+            embed.add_field(
+                name="Reprocessing",
+                value=content,
+                inline=(max_len < 25)
+            )
+        if is_loaded(item.compression) and len(item.compression) > 0:
+            content = "-# With Skill 000\n```\n"
+            max_len = max([len(comp.output.name) for comp in item.compression])
+            for comp in item.compression:
+                if comp.type == CompressionType.ore_low_density_compress:
+                    content += f"- {comp.output.name: <{max_len}}: {comp.artifice_num: <4} -> {comp.item_num}, via module\n"
+                else:
+                    content += f"  {comp.output.name: <{max_len}}: {comp.artifice_num: <4} -> {comp.item_num}, {comp.fuel_cost} GJ\n"
+            content += "```"
+            embed.add_field(
+                name="Compression",
+                value=content,
+                inline=False
+            )
+        return embed
